inventory_server/requirements.txt
uvicorn==0.20.0
python-decouple==3.7
fastapi==0.92.0
jose==1.0.0
passlib==1.7.4
pydantic==1.10.2
python_jose==3.3.0
requests==2.28.0
starlette==0.25.0
elasticsearch==8.7.0
kafka_python
requests==2.28.0



